refact: move serializeEntityKey method to common
