+++ go/internal/feast/onlinestore/redisonlinestore.go
@@ -7,8 +7,6 @@ import (
 	"errors"
 	"fmt"
 
-	//"os"
-	"sort"
 	"strconv"
 	"strings"
 
@@ -339,80 +337,6 @@ func buildRedisKey(project string, entityKey *types.EntityKey, entityKeySerializ
 	}
 	fullKey := append(*serKey, []byte(project)...)
 	return &fullKey, nil
-}
-
-func serializeEntityKey(entityKey *types.EntityKey, entityKeySerializationVersion int64) (*[]byte, error) {
-	// Serialize entity key to a bytestring so that it can be used as a lookup key in a hash table.
-
-	if entityKeySerializationVersion < 3 {
-		return nil, fmt.Errorf("Serialization of entity key with version < 3 is removed. Please use version 3 by setting entity_key_serialization_version=3. To reserializa your online store featrues refer -  https://github.com/feast-dev/feast/blob/master/docs/how-to-guides/entity-reserialization-of-from-v2-to-v3.md")
-	}
-
-	// Ensure that we have the right amount of join keys and entity values
-	if len(entityKey.JoinKeys) != len(entityKey.EntityValues) {
-		return nil, fmt.Errorf("the amount of join key names and entity values don't match: %s vs %s", entityKey.JoinKeys, entityKey.EntityValues)
-	}
-
-	// Make sure that join keys are sorted so that we have consistent key building
-	m := make(map[string]*types.Value)
-
-	for i := 0; i < len(entityKey.JoinKeys); i++ {
-		m[entityKey.JoinKeys[i]] = entityKey.EntityValues[i]
-	}
-
-	keys := make([]string, 0, len(m))
-	for k := range entityKey.JoinKeys {
-		keys = append(keys, entityKey.JoinKeys[k])
-	}
-	sort.Strings(keys)
-
-	// Build the key
-	length := 5 * len(keys)
-	bufferList := make([][]byte, length)
-
-	// For entityKeySerializationVersion 3 and above, we add the number of join keys
-	// as the first 4 bytes of the serialized key.
-	if entityKeySerializationVersion >= 3 {
-		byteBuffer := make([]byte, 4)
-		binary.LittleEndian.PutUint32(byteBuffer, uint32(len(keys)))
-		bufferList = append([][]byte{byteBuffer}, bufferList...)
-	}
-
-	for i := 0; i < len(keys); i++ {
-		offset := i * 2
-		byteBuffer := make([]byte, 4)
-		binary.LittleEndian.PutUint32(byteBuffer, uint32(types.ValueType_Enum_value["STRING"]))
-		bufferList[offset] = byteBuffer
-		bufferList[offset+1] = []byte(keys[i])
-	}
-
-	for i := 0; i < len(keys); i++ {
-		offset := (2 * len(keys)) + (i * 3)
-		value := m[keys[i]].GetVal()
-
-		valueBytes, valueTypeBytes, err := serializeValue(value, entityKeySerializationVersion)
-		if err != nil {
-			return valueBytes, err
-		}
-
-		typeBuffer := make([]byte, 4)
-		binary.LittleEndian.PutUint32(typeBuffer, uint32(valueTypeBytes))
-
-		lenBuffer := make([]byte, 4)
-		binary.LittleEndian.PutUint32(lenBuffer, uint32(len(*valueBytes)))
-
-		bufferList[offset+0] = typeBuffer
-		bufferList[offset+1] = lenBuffer
-		bufferList[offset+2] = *valueBytes
-	}
-
-	// Convert from an array of byte arrays to a single byte array
-	var entityKeyBuffer []byte
-	for i := 0; i < len(bufferList); i++ {
-		entityKeyBuffer = append(entityKeyBuffer, bufferList[i]...)
-	}
-
-	return &entityKeyBuffer, nil
 }
 
 func serializeValue(value interface{}, entityKeySerializationVersion int64) (*[]byte, types.ValueType_Enum, error) {
